requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
pyyaml==6.0
dvc==3.0.0
numpy==1.23.5
scipy==1.10.1
pygit2==1.12.2